move model & receiver to processors
